feat: 스프링 시큐리티 추가 및 유저 api 추가 (#11)

+++ src/main/kotlin/com/wheretopop/domain/user/UserInfoMapper.kt
@@ -11,7 +11,7 @@ class UserInfoMapper {
          */
         fun toMainInfo(user: User): UserInfo.Main {
             return UserInfo.Main(
-                id = user.id.toLong(),
+                id = user.id,
                 username = user.username,
                 email = user.email,
                 profileImageUrl = user.profileImageUrl,
